wip: setup resource tagging
