+++ api/pulumi/default/defaultLambdaRole.ts
