+++ api/pulumi/default/elasticSearch.ts
