+++ api/tsconfig.json
@@ -12,5 +12,5 @@
     "noImplicitReturns": true,
     "forceConsistentCasingInFileNames": true
   },
-  "files": ["pulumi/index.ts"]
+  "files": ["pulumi/**"]
 }
