+++ packages/cli-plugin-deploy-pulumi/utils/index.js
@@ -2,5 +2,6 @@ const getPulumi = require("./getPulumi");
 const getStackOutput = require("./getStackOutput");
 const mapStackOutput = require("./mapStackOutput");
 const loadEnvVariables = require("./loadEnvVariables");
+const tagResources = require("./tagResources");
 
-module.exports = { getPulumi, getStackOutput, loadEnvVariables, mapStackOutput };
+module.exports = { getPulumi, getStackOutput, loadEnvVariables, mapStackOutput, tagResources };
