+++ packages/pulumi-sdk/package.json
@@ -14,8 +14,8 @@
     "directory": "dist"
   },
   "dependencies": {
-    "@pulumi/aws": "3.22.0",
-    "@pulumi/pulumi": "2.19.0",
+    "@pulumi/aws": "^3",
+    "@pulumi/pulumi": "^2",
     "decompress": "^4.2.1",
     "download": "^5.0.3",
     "execa": "^4.0.3",
