+++ packages/pulumi-sdk/src/index.ts
@@ -32,13 +32,6 @@ export const FLAG_NON_INTERACTIVE = "--non-interactive";
 export const PULUMI_FOLDER = path.join(__dirname, "pulumi", os.platform());
 export const PULUMI_BINARY_PATH = path.join(PULUMI_FOLDER, "pulumi", "pulumi");
 
-// When updating versions, make sure to update it in three places:
-// - the AWS_PLUGIN_VERSION in this file, below this comment
-// – in `downloadBinaries.ts` file
-// - in `package.json`.
-// Check the README for more information.
-const AWS_PLUGIN_VERSION = "3.22.0";
-
 export class Pulumi {
     defaultArgs: DefaultArgs;
     constructor(options: DefaultArgs = {}) {
@@ -90,9 +83,10 @@ export class Pulumi {
         );
 
         if (installed) {
+            const { version } = require("@pulumi/aws/package.json");
             await execa(
                 path.join(PULUMI_FOLDER, "pulumi", "pulumi"),
-                ["plugin", "install", "resource", "aws", `v${AWS_PLUGIN_VERSION}`],
+                ["plugin", "install", "resource", "aws", `v${version}`],
                 {
                     stdio: "inherit",
                     env: {
